+++ tests/test_phylo_cluster.py
@@ -167,6 +167,9 @@ class PhyloClusterTester(unittest.TestCase):
                            "--assign_output", get_test_data("test_output_TarA"),
                            "--output", self.tmp_dir,
                            "--alpha", str(0.4)])
+        self.assertTrue(os.path.isfile(os.path.join(self.tmp_dir, "final_outputs", "phylotu_taxa.tsv")))
+        self.assertTrue(os.path.isfile(os.path.join(self.tmp_dir, "final_outputs", "phylotu_matrix.tsv")))
+        self.assertTrue(os.path.isfile(os.path.join(self.tmp_dir, "final_outputs", "phylotu_pquery_assignments.tsv")))
         return
 
     def test_dereplicate_by_clustering(self):